[refactor]: created input system
