+++ CMakeLists.txt
@@ -22,6 +22,8 @@ add_library(
   src/systems/camera_system.cpp
   src/systems/hud_system.h
   src/systems/hud_system.cpp
+  src/systems/input_system.h
+  src/systems/input_system.cpp
   dependencies/imgui/imgui.cpp
   dependencies/imgui/imgui_demo.cpp
   dependencies/imgui/imgui_draw.cpp
@@ -81,6 +83,8 @@ if(NOT BUILD_BENCHMARKS)
     src/systems/render_system.cpp
     src/systems/hud_system.h
     src/systems/hud_system.cpp
+    src/systems/input_system.h
+    src/systems/input_system.cpp
   )
 
   target_include_directories(fluid-sim PRIVATE
