+++ imgui.ini
@@ -12,5 +12,5 @@ Size=300,200
 
 [Window][##hud]
 Pos=20,20
-Size=319,213
+Size=107,98
