+++ src/systems/input_system.h
@@ -1,3 +1,16 @@
-#include <glad/glad.h>
-#include <GLFW/glfw3.h>
+#pragma once
+#include <imgui/imgui.h>
+#include <imgui/imgui_impl_glfw.h>
+#include <imgui/imgui_impl_opengl3.h>
 
+#include <algorithm>
+
+#include <GLFW/glfw3.h>
+#include "../app/app_state.h"
+#include "../linear_algebra.h"
+
+void ScrollCallback(GLFWwindow *window, double xOffset, double yOffset);
+void KeyCallback(GLFWwindow *window, int key, int scancode, int action,
+                 int mods);
+void MouseButtonCallback(GLFWwindow *window, int button, int action, int mods);
+void CursorPosCallback(GLFWwindow *window, double xpos, double ypos);
